backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
anthropic==0.7.8
twilio==8.10.0
sendgrid==6.11.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
